requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
drf-writable-nested==0.6.4
factory-boy==3.2.1
Faker==13.15.0
idna==3.3
iso-language-codes==1.1.0
model-bakery==1.6.0
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyodbc==4.0.34
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tmdbsimple==2.9.1
tzdata==2022.1
urllib3==1.26.10
